Add an option to ignore missing data
Technically the 'missing data' are not crucial for functioning, so crashing
completely might interfere with some uses. However, giving false peace of mind
would be frustrating too. Add a switch that defaults to the safest behaviour.

Also add a redundant comment in the example to alert users to its presence.